Front-end 5
Navigation bar is complete.
